backend/pyproject.toml
[tool.poetry]
name = "cosmos-api-index"
version = "0.1.0"
description = "Cosmos API - Index app"
authors = ["Maria Ibanez <mariaibanez@delosintelligence.fr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.1"
fastapi = "^0.115.6"
python-multipart = "^0.0.20"
delos-cosmos = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pydantic = "^2.10.5"
ruff = "^0.9.2"
pyright = "^1.1.392.post0"
mypy = "^1.14.1"

[tool.ruff]
line-length = 120
